ModyfingGenreIdInGigsClass

+++ gighub/Models/Gig.cs
@@ -24,7 +24,8 @@ namespace gighub.Models
 
 
         public Genre Genre { get; set; }
+
         [Required]
-        public int GenreId { get; set; }
+        public byte GenreId { get; set; }
     }
 }
